[topgen] Break out top RAL output into separate script
Mostly copying over the relevant functions from topgen into the new
Python script `util/gen_top_ral.py`.
